backend/requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-groq
fastapi
uvicorn